separate hastable trait
